# lib/api-zod/src/generated/types/addLeadNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddLeadNoteBody {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  note: string;
}