# lib/api-zod/src/generated/types/captureLeadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaptureLeadBody {
  /**
   * @minLength 2
   * @maxLength 200
   */
  name: string;
  email: string;
  /** @maxLength 30 */
  phone?: string;
  /** @maxLength 500 */
  address?: string;
  /**
   * @minLength 5
   * @maxLength 10
   */
  zipCode: string;
  diagnosisId?: string;
  /** @maxLength 1000 */
  notes?: string;
}