# lib/api-zod/src/generated/types/captureLeadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CaptureLeadResult {
  id: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  leadScore: number;
  matchedProfessional: boolean;
  message: string;
}