# lib/api-zod/src/generated/types/proLead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface ProLead {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  zipCode: string;
  status: LeadStatus;
  leadScore: number;
  /** @nullable */
  diagnosisId?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}