# lib/api-zod/src/generated/types/proLeadNote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProLeadNote {
  id: string;
  leadId: string;
  professionalId: string;
  note: string;
  createdAt: Date;
}