# lib/api-zod/src/generated/types/proLoginResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProProfileSummary } from "./proProfileSummary";

export interface ProLoginResult {
  token: string;
  professional: ProProfileSummary;
}