# lib/api-zod/src/generated/types/proProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProProfileSubscriptionStatus } from "./proProfileSubscriptionStatus";

export interface ProProfile {
  id: string;
  email: string;
  businessName: string;
  ownerName: string;
  /** @nullable */
  phone?: string | null;
  approved: boolean;
  subscriptionStatus: ProProfileSubscriptionStatus;
  serviceZipCodes: string[];
  servicesOffered: string[];
  /** @nullable */
  rating?: number | null;
  createdAt: Date;
}