# lib/api-zod/src/generated/types/proRegisterResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProRegisterResult {
  id: string;
  email: string;
  businessName: string;
  approved: boolean;
  message: string;
}