# lib/api-zod/src/generated/types/updateLeadStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface UpdateLeadStatusBody {
  status: LeadStatus;
}