# lib/api-zod/src/generated/types/updateLeadStatusResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LeadStatus } from "./leadStatus";

export interface UpdateLeadStatusResult {
  id: string;
  status: LeadStatus;
}