# lib/api-zod/src/generated/types/updateProProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LawnIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProProfileBody {
  /** @maxLength 30 */
  phone?: string;
  /**
   * @minItems 1
   * @maxItems 50
   */
  serviceZipCodes?: string[];
  /**
   * @minItems 1
   * @maxItems 20
   */
  servicesOffered?: string[];
}